# backend/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "CricZone API",
    "version": "1.0.0",
    "description": "Core authentication and cricket match endpoints. Additional controller coverage is tracked as a documentation follow-up.",
    "license": { "name": "ISC", "url": "https://opensource.org/license/isc-license-txt/" }
  },
  "servers": [
    {
      "url": "/",
      "description": "Current CricZone deployment"
    }
  ],
...
  "security": [],
  "paths": {
    "/api/users/register": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Register a user and create a session",
        "operationId": "registerUser",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RegisterInput" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User registered",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AuthResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "429": { "$ref": "#/components/responses/RateLimited" }
        }
      }
    },
    "/api/users/login": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Log in and create a session",
        "operationId": "loginUser",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginInput" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authenticated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AuthResponse" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "423": {
            "description": "Account temporarily locked",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "429": { "$ref": "#/components/responses/RateLimited" }
        }
      }
    },
    "/api/users/forgot-password": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Request a password reset email",
        "operationId": "forgotPassword",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ForgotPasswordInput" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Generic response returned for both registered and unregistered emails",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MessageResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "429": { "$ref": "#/components/responses/RateLimited" }
        }
      }
    },
    "/api/users/reset-password/{token}": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Set a new password using a time-limited reset token",
        "operationId": "resetPassword",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[a-fA-F0-9]{64}$"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ResetPasswordInput" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password reset and existing sessions revoked",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MessageResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "429": { "$ref": "#/components/responses/RateLimited" }
        }
      }
    },
    "/api/users/refresh": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Rotate a refresh session and issue a new access token",
        "operationId": "refreshSession",
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RefreshTokenInput" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Session refreshed",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/RefreshResponse" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/api/users/logout": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Revoke the current refresh session",
        "operationId": "logoutUser",
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RefreshTokenInput" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Logged out",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MessageResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
    "responses": {
      "BadRequest": {
        "description": "Invalid request",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "Unauthorized": {
        "description": "Authentication required or invalid",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
...
      "RateLimited": {
        "description": "Rate limit exceeded",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
...
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": ["success", "message"],
        "properties": {
          "success": { "type": "boolean", "example": false },
          "message": { "type": "string" },
          "requestId": { "type": "string" },
          "details": { "type": "array", "items": { "type": "object", "additionalProperties": true } }
        }
      },
      "MessageResponse": {
        "type": "object",
        "required": ["success", "message"],
        "properties": {
          "success": { "type": "boolean", "example": true },
          "message": { "type": "string" }
        },
        "additionalProperties": true
      },
      "RegisterInput": {
        "type": "object",
        "required": ["name", "email", "phone", "password"],
        "additionalProperties": false,
        "properties": {
          "name": { "type": "string", "minLength": 2, "maxLength": 80 },
          "email": { "type": "string", "format": "email", "maxLength": 254 },
          "phone": { "type": "string", "maxLength": 20 },
          "password": { "type": "string", "format": "password", "minLength": 6, "maxLength": 128 }
        }
      },
      "LoginInput": {
        "type": "object",
        "required": ["email", "password"],
        "additionalProperties": false,
        "properties": {
          "email": { "type": "string", "format": "email", "maxLength": 254 },
          "password": { "type": "string", "format": "password", "minLength": 1, "maxLength": 128 }
        }
      },
      "ForgotPasswordInput": {
        "type": "object",
        "required": ["email"],
        "additionalProperties": false,
        "properties": {
          "email": { "type": "string", "format": "email", "maxLength": 254 }
        }
      },
      "ResetPasswordInput": {
        "type": "object",
        "required": ["password", "confirmPassword"],
        "additionalProperties": false,
        "properties": {
          "password": { "type": "string", "format": "password", "minLength": 8, "maxLength": 128 },
          "confirmPassword": { "type": "string", "format": "password", "minLength": 8, "maxLength": 128 }
        }
      },
      "RefreshTokenInput": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "refreshToken": { "type": "string", "minLength": 32, "maxLength": 256 }
        },
        "description": "The refresh token may alternatively be supplied by the HttpOnly refreshToken cookie."
      },
      "User": {
        "type": "object",
        "properties": {
          "_id": { "type": "string" },
          "name": { "type": "string" },
          "email": { "type": "string", "format": "email" },
          "phone": { "type": "string" },
          "role": { "type": "string" }
        },
        "additionalProperties": true
      },
      "AuthResponse": {
        "type": "object",
        "required": ["success", "token", "expiresIn", "user"],
        "properties": {
          "success": { "type": "boolean", "example": true },
          "token": { "type": "string" },
          "expiresIn": { "type": "string", "example": "30m" },
          "user": { "$ref": "#/components/schemas/User" }
        },
        "additionalProperties": true
      },
      "RefreshResponse": {
        "type": "object",
        "required": ["success", "token", "expiresIn"],
        "properties": {
          "success": { "type": "boolean", "example": true },
          "token": { "type": "string" },
          "expiresIn": { "type": "string", "example": "30m" }
        }
      },
...
    }
  }
}